feat: add convert mod
