+++ crates/motiongfx_vello/src/vector_style.rs
@@ -1,27 +1,8 @@
+use crate::convert::*;
 use bevy_render::prelude::*;
 use bevy_utils::prelude::*;
 use bevy_vello_renderer::vello::{kurbo, peniko};
 use motiongfx_core::prelude::*;
-
-pub struct PenikoBrush(peniko::Brush);
-pub struct PenikoColor(peniko::Color);
-
-impl From<Color> for PenikoColor {
-    fn from(value: Color) -> Self {
-        PenikoColor(peniko::Color::rgba(
-            value.r() as f64,
-            value.g() as f64,
-            value.b() as f64,
-            value.a() as f64,
-        ))
-    }
-}
-
-impl From<Color> for PenikoBrush {
-    fn from(value: Color) -> Self {
-        PenikoBrush(peniko::Brush::Solid(PenikoColor::from(value).0))
-    }
-}
 
 #[derive(Clone)]
 pub struct FillStyle {
